feat: Use Constants instead of Strings
