+++ app/src/main/java/org/cosmic/ide/util/Constants.kt
@@ -5,4 +5,7 @@ object Constants {
     const val GITHUB_URL = "https://github.com/Cosmic-Ide/Cosmic-Ide"
     const val GITHUB_RELEASE_URL = "https://github.com/Cosmic-Ide/Cosmic-Ide/releases"
     const val DISCORD_URL = "https://discord.gg/zqaMxGeRyr"
+
+    const val PROJECT_PATH = "project_path"
+    const val DRAWER_STATE = "drawer_state"
 }
