+++ app/src/main/java/org/cosmic/ide/util/Extensions.kt
@@ -1,6 +1,24 @@
 package org.cosmic.ide.util
 
+import android.app.Activity
+import android.content.Context
+import androidx.annotation.StyleRes
+import androidx.core.app.ActivityCompat
 import androidx.lifecycle.LiveData
+
+private val getThemeResIdMethod by lazyReflectedMethod(Context::class.java, "getThemeResId")
+
+fun Activity.recreateCompat() {
+    ActivityCompat.recreate(this)
+}
+
+fun Activity.setThemeCompat(@StyleRes resid: Int) {
+    setTheme(resid)
+}
+
+val Context.themeResIdCompat: Int
+    @StyleRes
+    get() = getThemeResIdMethod.invoke(this) as Int
 
 @Suppress("UNCHECKED_CAST")
 val <T> LiveData<T>.valueCompat: T
